exposed interface for LibraryCache + added stub in CdkTestBase
