+++ generator/src/main/java/org/richfaces/cdk/Generator.java
@@ -36,6 +36,7 @@ import java.util.logging.Level;
 import org.richfaces.cdk.apt.AptModule;
 import org.richfaces.cdk.apt.CacheType;
 import org.richfaces.cdk.apt.LibraryCache;
+import org.richfaces.cdk.apt.LibraryCacheImpl;
 import org.richfaces.cdk.generate.java.ClassGeneratorModule;
 import org.richfaces.cdk.generate.taglib.TaglibModule;
 import org.richfaces.cdk.model.ModelModule;
@@ -180,7 +181,7 @@ public class Generator {
                 bind(FileManager.class).annotatedWith(new SourceImpl(entry.getKey())).toInstance(entry.getValue());
             }
             for (CacheType cacheType : CacheType.values()) {
-                LibraryCache cache = new LibraryCache(cacheType);
+                LibraryCache cache = new LibraryCacheImpl(cacheType);
                 requestInjection(cache);
                 bind(LibraryCache.class).annotatedWith(new CacheImpl(cacheType)).toInstance(cache);
             }
